+++ src/Data/Generics/Internal/GenericN.hs
@@ -23,7 +23,7 @@
 --------------------------------------------------------------------------------
 
 module Data.Generics.Internal.GenericN
-  ( Param
+  ( Param (..)
   , Rec (Rec, unRec)
   , GenericN (..)
   ) where
@@ -33,32 +33,23 @@ import GHC.Generics
 import GHC.TypeLits
 import Data.Coerce
 
-type family Param :: Nat -> k where
+data family Param :: Nat -> j -> k
+
+newtype instance Param n (a :: Type)
+  = StarParam { getStarParam :: a}
 
 type family Indexed (t :: k) (i :: Nat) :: k where
-  Indexed (t a) i = Indexed t (i + 1) (Param i)
+  Indexed (t a) i = Indexed t (i + 1) (Param i a)
   Indexed t _     = t
 
 newtype Rec (p :: Type) a x = Rec { unRec :: K1 R a x }
-
-type family Zip (a :: Type -> Type) (b :: Type -> Type) :: Type -> Type where
-  Zip (M1 mt m s) (M1 mt m t)
-    = M1 mt m (Zip s t)
-  Zip (l :+: r) (l' :+: r')
-    = Zip l l' :+: Zip r r'
-  Zip (l :*: r) (l' :*: r')
-    = Zip l l' :*: Zip r r'
-  Zip (Rec0 p) (Rec0 a)
-    = Rec p a
-  Zip U1 U1
-    = U1
 
 class
   ( Coercible (Rep a) (RepN a)
   , Generic a
   ) => GenericN (a :: Type) where
   type family RepN (a :: Type) :: Type -> Type
-  type instance RepN a = Zip (Rep (Indexed a 0)) (Rep a)
+  type instance RepN a = Rep (Indexed a 0)
   toN :: RepN a x -> a
   fromN :: a -> RepN a x